♻️ 함수를 utils-helperFunction 폴더로 분리
