+++ client/src/pages/Booking/components/BookingDates.tsx
@@ -6,22 +6,8 @@ import {
   ReservationDate,
   SeparationLine,
 } from '../style';
+import { makeDateFilledWithZero } from '../../../common/utils/helperFunctions/makeDateFilledWithZero';
 import { RootState } from '../../../common/store/RootStore';
-
-const makeDateFilledWithZero = (
-  reservationDate: { year: number; month: number; date: number } | null,
-) => {
-  const fillZero = (number: number) => {
-    return String(number).padStart(2, '0');
-  };
-
-  if (reservationDate) {
-    const { year, month, date } = reservationDate;
-    return `${year}.${fillZero(month)}.${fillZero(date)}`;
-  } else {
-    return null;
-  }
-};
 
 function BookingDates() {
   const reservationStartDate = useSelector(
